pyproject.toml
[tool.poetry]
name = "flask-ai-lawyer"
version = "0.1.0"
description = ""
authors = ["AndrewCheUA <andreicheban@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.1"
flask-login = "^0.6.3"
flask-jwt-extended = "^4.6.0"
flask-mail = "^0.9.1"
flask-bcrypt = "^1.0.1"
pillow = "^10.2.0"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
email-validator = "^2.1.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
